pyproject.toml
[project]
name = "sse-auto-translator"
version = "3.0.0-alpha-1"
description = "SSE Auto Translator"
license = { file = "LICENSE" }
authors = [{ name = "Cutleast", email = "cutleast@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4",
    "cloudscraper",
    "cutleast-core-lib",
    "deepl",
    "googletrans==3.1.0a0",
    "lingua-language-detector",
    "lz4",
    "plyvel-ci",
    "pytz",
    "pyuac",
    "pywin32",
    "pyzmq",
    "spylls",
    "sse-bsa",
    "websocket-client",
]

[dependency-groups]
dev = [
    "nuitka",
    "pyfakefs",
    "pyright",
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "pytest-qt",
    "ruff",
    "types-requests",
]

[tool.uv.workspace]
members = [
    "core-lib",
]

[tool.uv.sources]
cutleast-core-lib = { workspace = true }

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "[%(asctime)s.%(msecs)03d][%(levelname)s][%(name)s.%(funcName)s]: %(message)s"

[tool.pyright]
include = ["src", "tests"]
typeCheckingMode = "strict"
exclude = ["**/__pycache__", "**/*.old", "src/resources_rc.py"]
stubPath = "./stubs"

# Check here for a full list of options: https://github.com/microsoft/pyright/blob/main/docs/configuration.md#diagnostic-settings-defaults
deprecateTypingAliases = false
reportImplicitOverride = "warning"
reportMatchNotExhaustive = false
reportMissingTypeArgument = false
reportMissingTypeStubs = false
reportPropertyTypeMismatch = "warning"
reportShadowedImports = "warning"
reportUnknownArgumentType = false
reportUnknownLambdaType = false
reportUnknownMemberType = false
reportUnknownParameterType = false
reportUnknownVariableType = false
reportUnnecessaryCast = "information"
reportUnnecessaryComparison = "information"
reportUnnecessaryContains = "information"
reportUnnecessaryIsInstance = "information"
reportUnnecessaryTypeIgnoreComment = "information"
reportUnusedClass = "warning"
reportUnusedImport = "warning"
reportUnusedFunction = "warning"
reportUnusedVariable = "warning"
